order tests

+++ src/main/java/com/example/webbshopbackend1/repositories/OrderRepository.java
@@ -4,7 +4,7 @@ import com.example.webbshopbackend1.models.Customer;
 import com.example.webbshopbackend1.models.Orders;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Orders, Integer> {
+public interface OrderRepository extends CrudRepository<Orders, Long> {
 
     Iterable<Orders> findByCustomer(Customer customer);
